See description
This is mainly a description of the last update- I changed the way the
file system was structured to increase abstraction. Basically, I split
things into two packages- one with the MVC stuff, one with the old
stuff. If you're not sure how packages effect methods, check this out-
https://docs.oracle.com/javase/tutorial/java/javaOO/accesscontrol.html"
I also changed the model so that it's now a model of the Country class,
as most of the console stuff we did with TeamMateModel is not needed by
this program.
If you have more questions, let me know- I'll be updating the UML
diagram shortly to help reflect the changes that have been made.